Picodrive: Implement more mapper from libretro Update translations and bump version number Fix warnings Add hack for unlicensed games that don't handle the Z80 bus properly Update buildin carthw.cfg

+++ src/cores/picodrive/pico/pico.c
@@ -169,7 +169,7 @@ int PicoReset(void)
  
   // bug fix? without this, the jump table for the PicoCpuCS68k's jumptab is never initialized
   // and the CPU slave emulation will crash when loading a state before the CD is initialized.
-  PicoCpuCS68k.jumptab = &CycloneJumpTab;
+  PicoCpuCS68k.jumptab = (unsigned int) (long) &CycloneJumpTab;
 
   // ..but do not reset SekCycle* to not desync with addons
